优化了界面，但第三方框架导入有些问题，图片加载问题还没解决

+++ RuiSi/RuiSi/Section/View/SectionCollectionViewCell.m
@@ -23,7 +23,9 @@
 
 - (void)configureImageViewForIndexPath:(NSIndexPath *)indexPath {
     
-    NSString *fileName = [NSString stringWithFormat:@"%ld_%ld.png",(long)(indexPath.section+1),(long)(indexPath.row+1)];
+    NSString *fileName = [NSString stringWithFormat:@"%ld_%ld.png",(indexPath.section+1),(indexPath.row+1)];
+    
+   // NSLog(@"%ld_%ld.png",(indexPath.section+1),(indexPath.row+1));
     
     [self.iconImageView setImage:[UIImage imageNamed:fileName]];
 }
